refactor: ♻️ repurpose Cli class to Configs
Renamed Cli to ConfigsProcessor, now the Cli class is purposed to just
parsing the cli arguments and orchestrate other classes
